mcp-api-hub/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-api-hub"
version = "0.1.0"
description = "Generalized MCP server exposing multiple API tools via the Model Context Protocol"
authors = [{ name = "Internal", email = "devnull@example.com" }]
requires-python = ">=3.9"
dependencies = [
  "mcp>=1.2.0",
  "pydantic>=2.5",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "tenacity>=8.2",
  "structlog>=24.1",
]
readme = "README.md"

[project.scripts]
# CLI entrypoint `mcp-api-hub` -> mcp_api_hub.server:main
mcp-api-hub = "mcp_api_hub.server:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["mcp_api_hub*"]
